// src/app/generated/openapi/notip-data-api-openapi/model/sensor-dto.ts
/**
 * NoTIP Data API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface SensorDto {
  /**
   * Sensor identifier
   */
  sensorId: string;
  /**
   * Sensor type
   */
  sensorType: string;
  /**
   * Gateway identifier
   */
  gatewayId: string;
  /**
   * Timestamp of the latest measure seen for the sensor
   */
  lastSeen: string;
}